+++ player-analytics-core/src/main/java/video/api/player/analytics/core/utils/PlayerAnalyticsUserAgent.kt
@@ -1,0 +1,7 @@
+package video.api.player.analytics.core.utils
+
+internal object PlayerAnalyticsUserAgent {
+    fun create(): String? {
+        return System.getProperty("http.agent")
+    }
+}