+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All changes to this project will be documented in this file.
 
+## [3.0.0] - 2024-07-24
+
+- Use Analytics endpoint v2
+- Use Kotlin DSL for Gradle
+
 ## [2.0.2] - 2023-10-09
 
 - Fix a crash on ExoPlayer Analytics when position < 0.0f